Asher Chen commited at 2023-01-15_07-01-31

+++ test/distances.cpp
@@ -13,7 +13,7 @@ using namespace graph::extra::io;
 
 using us_graph =
   graph::simple_graph<
-    uint16_t,
+    uint32_t,
     uint32_t,
     graph::tag::undirected
   >;
